OF-2272: Replace Local- and Remote- MUC entities with one generic implementation.
An issue arises from having a concept of “local” and “remote” MUC users and roles. In this context, ‘local’ refers to a entity that is connected to the local cluster node, while “remote” refers to an entity connected to another node in the cluster. This distinction cannot be made for entities that join a MUC room from remote domains (through server-to-server functionality). Server-to-server connectivity can arbitrarily be established to any one, but also multiple, nodes of the cluster. Worse, a server-to-server connection can be teared down, and re-established to a new cluster node. Because of this, it’s not possible to uniquely qualify a MUC user that is a user from another XMPP domain as a ‘remote’ or ‘local’ MUC user. It can be either, or both, but also change over time.

This commit broadly removes the concept of having “local” and “remote” entities, replacing them with a singular representation.

More work is needed to properly maintain a singular collection of entities in a shared memory datastructure.