OF-2220: Ensure MUCUser is updated in cache when changed
MUCUser instances are shared in a clustered data structure (a Cache) to allow all cluster nodes to work with the instances.

Whenever an instance is modified, it needs to explicitly put in the Cache (again), for the change to become visible to other cluster nodes.

This commit ensures that whenever the internal state of a MUCUser object is modified, it's being re-added to the cache. Each operation is protected by a cluster-wide lock for good measure.