OF-2272: Do not keep MUCRoles in a distinct cache
The migration from Remote- and LocalMUC entities to Cacheable entities introduced a cache for MUCRoles. As these instances are always very specific to exactly one MUCRoom, which already is in a Cache itself, the corresponding MUCRoles can be stored in a field of a MUCRoom, and be shared across a cluster as part of the room that they belong to.

The benefit of this is that the complexity of the solution significantly reduces. There's a lot less room for discrepancies between the state of different classes, and applying mutexes to prevent concurrent modification of cached entities could, in theory, be as simple as: lock the room whenever it gets a stanza to be processed (or is modified through another process, like a timer task or the admin console). There no longer is a need to lock different caches, which easily leads to deadlocks, if multiple cluster nodes each hold a different lock, and are waiting to acquire a lock currently held by another node.