Rebase onto v4.9.1 of OF-2904: Fix parsing issue of OtherName
OtherName values are tagged, which should be taken into account by the parser.

Unit tests that generate a certificate should include an EXPLIT tagged OtherName.

After two days of ... pain ... I've resorted to converting the BouncyCastle-generated certificate to PEM and back again, which for some reason that's beyond me appears to be required to make the parsing be successfull on both Java 17 and 21.

Adds additional unit test that uses a certificate generated by OpenSSL.

Many thanks to @dwd for invaluable help on this.