OF-2811: Must not block on eventloop
OF-2808 describes an issue in which a session that is resumed gets closed. This is caused by the asynchronous 'close' operation to execute 'connection close' listeners on a session that, in another thread, has already been resumed.

The fix for OF-2808 introduces a Latch, that causes a thread to block. If that thread is one of a Netty EventLoop, all processing stops (possibly leading to the latch never be released, as the connection close listeners cannot operate).

This commit reverts the change for OF-2808, and applies a new fix for that issue: disallow a session that's already in the process of being closed from resumed.